requirements.txt
langchain>=0.3.0
langchain-groq>=0.1.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4
chromadb>=0.5.5
sentence-transformers>=3.0.1
streamlit>=1.38.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
pydantic>=2.8.2
pandas>=2.2.3
numpy>=2.0.0
